Move DynamicMemory to shale and add tests
